+++ app/client/keybase/debug/keystore.go
@@ -2,12 +2,10 @@ package debug
 
 import (
 	"fmt"
-	"github.com/pokt-network/pocket/shared/converters"
 	"os"
-	"path/filepath"
-	r "runtime"
 
 	"github.com/pokt-network/pocket/app/client/keybase"
+	"github.com/pokt-network/pocket/build"
 	"github.com/pokt-network/pocket/logger"
 	"github.com/pokt-network/pocket/runtime"
 	cryptoPocket "github.com/pokt-network/pocket/shared/crypto"
@@ -19,9 +17,8 @@ import (
 
 const (
 	// NOTE: This is the number of validators in the private-keys.yaml manifest file
-	numValidators       = 999
-	debugKeybaseSuffix  = "/.pocket/keys"
-	privateKeysYamlFile = "../../../../../build/localnet/manifests/private-keys.yaml"
+	numValidators      = 999
+	debugKeybaseSuffix = "/.pocket/keys"
 )
 
 var (
@@ -131,20 +128,8 @@ func fetchValidatorPrivateKeysFromK8S() (map[string]string, error) {
 }
 
 func fetchValidatorPrivateKeysFromFile() (map[string]string, error) {
-	// BUG: When running the CLI using the build binary (i.e. `p1`), it searched for the private-keys.yaml file in `github.com/pokt-network/pocket/build/localnet/manifests/private-keys.yaml`
-	// Get private keys from manifest file
-	_, current, _, _ := r.Caller(0)
-	//nolint:gocritic // Use path to find private-keys yaml file from being called in any location in the repo
-	yamlFile := filepath.Join(current, privateKeysYamlFile)
-	if exists, err := converters.FileExists(yamlFile); !exists || err != nil {
-		return nil, fmt.Errorf("unable to find YAML file: %s", yamlFile)
-	}
 
 	// Parse the YAML file and load into the config struct
-	yamlData, err := os.ReadFile(yamlFile)
-	if err != nil {
-		return nil, err
-	}
 	var config struct {
 		ApiVersion string            `yaml:"apiVersion"`
 		Kind       string            `yaml:"kind"`
@@ -152,7 +137,7 @@ func fetchValidatorPrivateKeysFromFile() (map[string]string, error) {
 		Type       string            `yaml:"type"`
 		StringData map[string]string `yaml:"stringData"`
 	}
-	if err := yaml.Unmarshal(yamlData, &config); err != nil {
+	if err := yaml.Unmarshal(build.PrivateKeysFile, &config); err != nil {
 		return nil, err
 	}
 	validatorKeysMap := make(map[string]string)