+++ app/client/doc/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.19] - 2023-02-28
+
+- Introduces vault keybase to allow for the use of a vault server to store the private keys
+- Fixes bug where private keys were not being properly parsed from the debug keybase
+
 ## [0.0.0.18] - 2023-02-28
 
 - Implement SLIP-0010 HD child key derivation with the keybase